+++ sync-for-payables/CodatSyncPayables/Models/Operations/UnlinkConnectionRequest.cs
@@ -11,13 +11,12 @@
 namespace CodatSyncPayables.Models.Operations
 {
     using CodatSyncPayables.Utils;
-    using System.Collections.Generic;
     
     public class UnlinkConnectionRequest
     {
 
         [SpeakeasyMetadata("request:mediaType=application/json")]
-        public Dictionary<string, object>? RequestBody { get; set; }
+        public UnlinkConnectionUpdateConnection? RequestBody { get; set; }
 
         /// <summary>
         /// Unique identifier for a company.