+++ sync-for-payables/CodatSyncPayables/Models/Shared/PushOperationRef.cs
@@ -11,18 +11,20 @@
 namespace CodatSyncPayables.Models.Shared
 {
     using Newtonsoft.Json;
-    using System.Collections.Generic;
     
-    /// <summary>
-    /// A human-readable object describing validation decisions Codat has made when pushing data into the platform. If a push has failed because of validation errors, they will be detailed here.
-    /// </summary>
-    public class Validation
+    public class PushOperationRef
     {
 
-        [JsonProperty("errors")]
-        public List<ValidationItem>? Errors { get; set; }
+        /// <summary>
+        /// Available Data types
+        /// </summary>
+        [JsonProperty("dataType")]
+        public DataType? DataType { get; set; }
 
-        [JsonProperty("warnings")]
-        public List<ValidationItem>? Warnings { get; set; }
+        /// <summary>
+        /// Unique identifier for a push operation.
+        /// </summary>
+        [JsonProperty("id")]
+        public string? Id { get; set; }
     }
 }